requirements.txt
fastapi
uvicorn
boto3
elasticsearch>=8,<9
pydantic
python-dotenv
pytest
httpx
requests
typing_extensions
pyyaml
